Added a new "added by" to the tracker (#9)
This will let the tracker easily tell where this data has been added from without requiring the need to use the key to determine it. Which will help us decide if we need a popup dialog or bring the user to the info page instead when clicking on the item in the info page.

Took 4 minutes