Revamp how checking if ascension tier is tracked is obtained (#10)
Previously it will make a firebase call for every tier, this is extremely expensive and unnecessary to make these many calls. The change makes it such that instead of doing so, we now simply do a single firebase call to retrieve the list and compare it locally. This saves on calls as well as network use

Took 2 minutes