Moved weapons to unreleased weapons
This is because they were available in CBT but not the current live version of the game yet. Hence we remove it in case data is updated before they are re-released

Note that there are actually more weapons that are found but not released. We would not be including them here for now as that is too many things to keep track of

Removed Weapons
- Amber Catalyst
- Ebony Bow
- Quartz