build(dart): Upgraded all Firebase Dependencies
It is currently locked behind firebase_image dependency, which seems to be no longer maintained and we are considering replacing it in #616. However in the meantime we make use of an alternative to ensure that we are able to update the core firebase dependencies.

This also allows us to use the latest app_check dependency which allows for Android Debug attestation

Took 30 minutes