feat: Updated load file implementation to be more platform agnostic
We are using a seperate package to load file that supports loading from all platforms. Unfortunately for saving, files cannot be saved with this package hence we have to make do with different packages and custom code. We will remove it when this package ever gets updated to support it

Took 10 minutes